+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - #49 - entities number check after update tests
 - #78 - Generate test data for BankCardRestControllerIT
 - #87 - Generate test data for WorkingScheduleRestControllerIT
+- #35 - New ProductDto fields
 
 ## [1.0.0] - 30.07.2023
 - #1 - added working schedule entity